fix:修复kafka集群重新启动之后cluster.id不匹配的问题，因为docker中zookeeper没有通过volume持久化
