+++ docker-compose.yaml
@@ -87,6 +87,10 @@ services:
     ports:
       - "2181:2181"
     restart: always
+    volumes:
+      # 如果之后zookeeper版本升级了，可能需要修改这个路径
+      - ./data/zookeeper/data:/opt/zookeeper-3.4.13/data
+      - ./data/zookeeper/datalog:/opt/zookeeper-3.4.13/datalog
   kafka1:
     image: wurstmeister/kafka
     depends_on: [ zookeeper ]
@@ -96,6 +100,7 @@ services:
     environment:
       HOSTNAME: kafka1
 #      KAFKA_BROKER_ID: 0
+#      KAFKA_CLUSTER_ID: your_cluster_id
       KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS: 6000
       KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka1:9091
       KAFKA_LISTENERS: PLAINTEXT://0.0.0.0:9091
@@ -113,6 +118,7 @@ services:
     environment:
       HOSTNAME: kafka2
 #      KAFKA_BROKER_ID: 1
+#      KAFKA_CLUSTER_ID: your_cluster_id
       KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS: 6000
       KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka2:9092
       KAFKA_LISTENERS: PLAINTEXT://0.0.0.0:9092
@@ -130,6 +136,7 @@ services:
     environment:
       HOSTNAME: kafka3
 #      KAFKA_BROKER_ID: 2
+#      KAFKA_CLUSTER_ID: your_cluster_id
       KAFKA_ZOOKEEPER_CONNECTION_TIMEOUT_MS: 6000
       KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka3:9093
       KAFKA_LISTENERS: PLAINTEXT://0.0.0.0:9093
