+++ internal/order/service/redis.go
@@ -50,13 +50,26 @@ func ToDelayQueue(ctx context.Context, orderInfo string, timeUnix float64) {
 // RemoveFromDelayQueue 按时支付订单，从延时队列中取出该订单
 func RemoveFromDelayQueue(ctx context.Context, req *order.CommitOrderRequest) error {
 	//TODO 修改逻辑，保证幂等 1. 先查看redis中有没有，如果没有，说明订单已经过期，直接返回；2. 调用支付接口；3. 若用户成功支付，将order信息从zSet中删除并修改数据库、若用户支付失败，则不进行任何操作
-	// 第三步redis和MySQL不能保证原子性，这里我们加一个事务，只要保证MySQL状态更新成功就可以了。
+	// 第三步redis和MySQL不能保证原子性，这里我们只要保证MySQL状态更新成功就可以了。
 	// 因为就算redis删除失败，在之后订单自动过期的逻辑里，我们也会查看MySQL中订单状态，如果已经支付成功。是不会让订单被退回的。
 	orderInfo := fmt.Sprintf("%d;%d;%d;%d", req.UserId, req.ScheduleId, req.SeatRow, req.SeatCol)
-	log.Println("orderInfo = ", orderInfo)
-	count, err := redisC.ZRem(ctx, delayQueue, orderInfo).Result()
-	log.Println("count = ", count, " err = ", err)
+	logrus.Debug("orderInfo = ", orderInfo)
+	score, err := redisC.ZScore(ctx, delayQueue, orderInfo).Result()
+	logrus.Debug("score = ", score, " err = ", err)
 	if err != nil {
+		logrus.Error("failed to get order from delay queue:", err)
+		return err
+	}
+	// 模拟调用第三方支付接口，指定支付过期时间
+	time.Sleep(time.Second)
+	if time.Now().Unix() > int64(score) { //或者第三方接口返回支付过期，支付失败
+		return errors.New("订单已经过期")
+	}
+	// 按时支付成功，从延迟队列中删除订单
+	count, err := redisC.ZRem(ctx, delayQueue, orderInfo).Result()
+	logrus.Debug("count = ", count, " err = ", err)
+	if err != nil {
+		logrus.Error("failed to remove order from delay queue:", err)
 		return err
 	}
 	if count == 0 {
@@ -68,17 +81,16 @@ func RemoveFromDelayQueue(ctx context.Context, req *order.CommitOrderRequest) er
 		log.Println(err)
 		return err
 	}
-	// TODO 票的状态似乎并不需要保存是否付款，之后可以干掉这个逻辑
 	return nil
 }
 
 // toTargetQueue 处理延迟队列取出来的订单
 func toTargetQueue(ctx context.Context) {
 	for {
-		// 从延迟队列中取出最小的时间戳
+		// 从延迟队列中取出最小的时间戳，延迟3s,避免极端情况下，用户卡点支付，导致订单被错误地取消
 		result, err := redisC.ZRangeByScoreWithScores(ctx, delayQueue, &redis.ZRangeBy{
 			Min:    "-inf",
-			Max:    fmt.Sprintf("%d", time.Now().Unix()),
+			Max:    fmt.Sprintf("%d", time.Now().Add(time.Second*3).Unix()),
 			Offset: 0,
 			Count:  100,
 		}).Result()
@@ -119,7 +131,7 @@ func eventLoop(ctx context.Context) {
 		if errors.Is(err, redis.Nil) {
 			continue
 		}
-		logrus.Debug("now = ", time.Now().Format("2006-01-02 15:04:05"))
+		logrus.Debug("now = ", time.Now().Format(time.DateTime))
 		logrus.Debug("results = ", results)
 		if len(results) == 0 {
 			continue
@@ -127,7 +139,6 @@ func eventLoop(ctx context.Context) {
 		results = results[1:]
 		logrus.Debug("results = ", results)
 		for _, result := range results {
-			//TODO 为了防止订单在支付过程中，到达了过期时间，这里被意外退票。需要在这里判断一下订单状态是否为已支付，保证让已经支付的订单不会被抛弃。
 			fmt.Println("result = ", result)
 			data := strings.Split(result, ";")
 			fmt.Println("data = ", data)
@@ -143,24 +154,5 @@ func eventLoop(ctx context.Context) {
 			}
 		}
 
-	}
-}
-
-func sendMessage(msg kafka.Message) {
-	w := &kafka.Writer{
-		Addr:         kafka.TCP(consts.KafkaBroker1Url, consts.KafkaBroker2Url, consts.KafkaBroker3Url),
-		Topic:        consts.TicketTimeoutTopic,
-		Balancer:     &kafka.LeastBytes{}, //Hash, RoundRobin, LeastBytes
-		RequiredAcks: kafka.RequireOne,    //Leader写入成功即可
-	}
-	defer func() {
-		if err := w.Close(); err != nil {
-			logrus.Error("failed to close writer:", err)
-		}
-	}()
-
-	err := w.WriteMessages(context.Background(), msg)
-	if err != nil {
-		logrus.Error(w.Topic, "failed to write messages:", err)
 	}
 }
